6. ResponseEntity. Статусы ответов. [#505111]

+++ src/main/java/ru/job4j/socialmedia/controller/user/UserRestController.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import ru.job4j.socialmedia.model.User;
 import ru.job4j.socialmedia.service.user.UserService;
 
@@ -23,18 +24,29 @@ public class UserRestController {
 
     @PostMapping
     public ResponseEntity<User> createUser(@RequestBody User user) {
-        return ResponseEntity.ok(userService.createUser(user));
+        userService.createUser(user);
+        var uri = ServletUriComponentsBuilder
+                .fromCurrentRequest()
+                .path("/{id}")
+                .buildAndExpand(user.getId())
+                .toUri();
+        return ResponseEntity.created(uri).build();
     }
 
     @DeleteMapping("/{userId}")
-    @ResponseStatus(HttpStatus.OK)
-    public void deleteUserById(@PathVariable("userId") Integer id) {
+    public ResponseEntity<Void> deleteUserById(@PathVariable("userId") Integer id) {
+        if (userService.findById(id).isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
         userService.deleteUserById(id);
+        return ResponseEntity.noContent().build();
     }
 
     @PutMapping
-    @ResponseStatus(HttpStatus.OK)
-    public void updateUser(@RequestBody User user) {
-        userService.createUser(user);
+    public ResponseEntity<Void> updateUser(@RequestBody User user) {
+        if (userService.createUser(user) != null) {
+            return ResponseEntity.ok().build();
+        }
+        return ResponseEntity.notFound().build();
     }
 }
